feat: reorganize Nix installation script and update zshrc loading 🛠️
This commit refines the Nix installation process by replacing the old installer script with a more robust version, and it also updates the zsh configuration to improve clarity and functionality.

- Updated `zshrc` to format the conditional source statement for improved readability. This makes the configuration cleaner and easier to maintain.
- Deleted the old `install-nix` script which was less informative and less flexible.
- Introduced a new `nix-install` script that includes better logging and a more user-friendly installation process, allowing users to confirm if they wish to proceed with the installation if Nix is already detected.
- Removed the `nixd` script, which was not being used effectively, simplifying the repository and reducing clutter.

These changes make it easier for users to install Nix and work with it in a clearer and more manageable way.